pyproject.toml
[project]
name = "mylana"
version = "0.1.0"
description = "Personal finance AI agent system"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "crewai[tools]",
    "ddgs>=9.10.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.128.0",
    "langchain-community>=0.4.1",
    "litellm[proxy]>=1.75.3",
    "python-dotenv>=1.1.1",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mylana"]

[project.scripts]
mylana = "mylana.main:run"
run_crew = "mylana.main:run"
